requirements.txt
xmltodict~=0.12.0
discord.py~=1.5.1